requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
DiscordUtils==1.3.4
ffmpeg==1.4
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
mutagen==1.47.0
pycryptodomex==3.20.0
python-dotenv==1.0.1
pytube==15.0.0
requests==2.31.0
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
youtube-dl==2021.12.17
